Move public methods into jewelpet.github
